+++ src/model/dataset.ts
@@ -56,7 +56,12 @@ export interface TimeDimension extends Dimension {
 export type NormRange = [number, number];
 
 export interface RgbSchema {
-    tileSourceOptions: TileSourceOptions;
+    // The following are new since xcube 0.11
+    tileUrl: string;
+    tileLevelMin?: number;
+    tileLevelMax?: number;
+    // tileSourceOptions are longer used since xcube 0.11
+    tileSourceOptions?: TileSourceOptions;
     varNames: [string, string, string];
     normRanges: [NormRange, NormRange, NormRange];
 }